Rename osgifeature to osgi
