+++ lib/modules/manager/api.ts
@@ -55,7 +55,7 @@ import * as nodenv from './nodenv';
 import * as npm from './npm';
 import * as nuget from './nuget';
 import * as nvm from './nvm';
-import * as osgifeature from './osgifeature';
+import * as osgi from './osgi';
 import * as pipCompile from './pip-compile';
 import * as pip_requirements from './pip_requirements';
 import * as pip_setup from './pip_setup';
@@ -141,7 +141,7 @@ api.set('nodenv', nodenv);
 api.set('npm', npm);
 api.set('nuget', nuget);
 api.set('nvm', nvm);
-api.set('osgifeature', osgifeature);
+api.set('osgi', osgi);
 api.set('pip-compile', pipCompile);
 api.set('pip_requirements', pip_requirements);
 api.set('pip_setup', pip_setup);
