+++ lib/modules/manager/osgi/__fixtures__/bundles-as-objects.json
