+++ lib/modules/manager/osgi/__fixtures__/double-slash-not-comment.json
