+++ lib/modules/manager/osgi/__fixtures__/extension-artifacts.json
