+++ lib/modules/manager/osgi/__fixtures__/framework-artifact.json
