+++ lib/modules/manager/osgi/__fixtures__/unsupported-feature-version.json
