+++ lib/modules/manager/osgi/__fixtures__/various-malformed-definitions.json
