+++ lib/modules/manager/osgi/__fixtures__/version-with-variable.json
