+++ lib/modules/manager/osgi/__fixtures__/with-comment.json
