+++ lib/modules/manager/osgi/index.ts
