+++ lib/modules/manager/osgi/types.ts
